+++ src/Sentry/Internal/ScopeObserver.cs
@@ -65,7 +65,7 @@ internal abstract class ScopeObserver : Sentry.IScopeObserver
 
     public abstract void UnsetTagImpl(string key);
 
-    public void SetUser(User? user)
+    public void SetUser(SentryUser? user)
     {
         if (user is null)
         {
@@ -82,7 +82,7 @@ internal abstract class ScopeObserver : Sentry.IScopeObserver
         }
     }
 
-    public abstract void SetUserImpl(User user);
+    public abstract void SetUserImpl(SentryUser user);
 
     public abstract void UnsetUserImpl();
 }